moved App to components
